Init password tests and split in separate test files
