+++ packages/strapi-plugin-content-manager/test/groups.test.e2e.js
@@ -1,0 +1,3 @@
+'use strict';
+
+describe('Testing groups CRUD', () => {});
