+++ packages/strapi-plugin-content-manager/test/types/email.test.e2e.js
@@ -1,0 +1,7 @@
+describe('Test type email', () => {
+  test.todo('Create entry with valid value');
+
+  test.todo('Reading entry, returns correct value');
+
+  test.todo('Updating entry sets the right value and format');
+});
